// web/src/lib/api/generated/models/githubComViewraViewraInternalApplicationTvTVEpisodeResponse.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * ViewRA Media Server API
 * Self-hosted media server for movies, TV shows, and music
 * OpenAPI spec version: 0.0.1
 */

export interface GithubComViewraViewraInternalApplicationTvTVEpisodeResponse {
  air_date?: string
  audio_codec?: string
  bitrate?: number
  container_format?: string
  created_at?: string
  description?: string
  /** in seconds */
  duration?: number
  episode?: number
  /** actual episode name */
  episode_title?: string
  file_path?: string
  file_size?: number
  frame_rate?: number
  height?: number
  /** Base Media fields */
  id?: number
  imdb_id?: string
  is_extra?: boolean
  library_id?: number
  season?: number
  season_id?: number
  /** TV Episode-specific fields */
  show_title?: string
  /** filename */
  title?: string
  tvdb_id?: number
  updated_at?: string
  video_codec?: string
  /** Technical metadata */
  width?: number
}